fix(agent-builder): move useResolvedMessageContent and useReplacementsHold to hooks
