+++ x-pack/platform/plugins/shared/agent_builder/public/application/hooks/use_replacements_hold.ts
