+++ x-pack/platform/plugins/shared/agent_builder/public/application/hooks/use_resolved_message_content.test.ts
